kernel/futex: compare FUTEX_WAIT/CMP_REQUEUE value as 32-bit (fix int sign-extension EAGAIN storm) (#489)
Per futex(2), "the futex word is a 32-bit field" and the FUTEX_WAIT
comparison `*uaddr == val` is defined over those 32 bits only. Our
check-then-queue helper compared the full 64-bit `val` syscall argument
against the zero-extended 32-bit `*uaddr`, which breaks for any futex
word with bit 31 set.

The futex `val` argument is typed `int` in the C-library wrappers
(pthread_mutex_timedlock -> __timedwait(addr, val, ...)). When a thread
waits on a contended mutex, musl passes the mutex word with the
waiters bit (0x8000_0010) as that `int`. Widening a negative `int` to
the 64-bit syscall register sign-extends it, so the kernel receives
0xFFFF_FFFF_8000_0010 for a futex word of 0x8000_0010. The full-width
compare `current as u64 != val` then mismatched in the high half on
every call and returned EAGAIN, so a waiter on a contended mutex spun
instead of parking — observed as a multi-billion-call FUTEX_WAIT storm
with a frozen futex word, and a main thread that never slept to receive
the owner's release wake.

Fix: truncate `val` to 32 bits before the compare (`current != val as u32`),
matching the 32-bit futex-word contract. The same defect existed in the
FUTEX_CMP_REQUEUE `*uaddr == val3` check (pthread_cond broadcast path);
fixed there too.

Verified against a real workload that drove a contended-mutex wait whose
word's bit 31 was set: pre-fix the syscall count climbed past 4x10^8 with
the futex word frozen and no waiter ever enqueued; post-fix the same
workload's syscall count stays bounded (max ~2.8x10^3), waiters enqueue
and park correctly, and execution proceeds past the spin. Reproduced
across two runs with identical outcome.

Adds a deterministic test_runner.rs regression test
(test_futex_wait_val_is_32bit) that drives futex_wait_check_and_enqueue
with a sign-extended `val` and asserts: low-32-bits-agree -> Enqueued,
low-32-bits-differ -> ValueMismatch.

Refs:
  * futex(2) — FUTEX_WAIT / FUTEX_CMP_REQUEUE value comparison, 32-bit
    futex word.
  * POSIX-1.2017 pthread_mutex_lock(3) / pthread_mutex_timedlock(3).

+++ kernel/src/syscall/mod.rs
@@ -374,7 +374,21 @@ where
         Some(v) => v,
         None => return FutexWaitOutcome::Fault,
     };
-    if current as u64 != val {
+    // Per futex(2): "The futex word is a 32-bit field."  The FUTEX_WAIT
+    // comparison is `*uaddr == val` over those 32 bits only — the kernel
+    // must ignore any bits the caller placed above bit 31 of the `val`
+    // syscall argument.  This matters because the futex `val` argument is
+    // typed `int` in the C library wrappers (e.g. pthread_mutex_timedlock's
+    // `__timedwait(addr, val, …)` where `val` carries the mutex word with
+    // bit 31 — the contended-waiters bit — set).  Widening a negative `int`
+    // to the 64-bit syscall register sign-extends it: a futex word of
+    // 0x8000_0010 arrives as 0xFFFF_FFFF_8000_0010.  Comparing the
+    // zero-extended 32-bit `*uaddr` (0x0000_0000_8000_0010) against the full
+    // 64-bit `val` would then mismatch in the high half on every call,
+    // returning EAGAIN forever and never parking the waiter — a contended
+    // mutex's owner-handoff wait spins instead of sleeping.  Truncate `val`
+    // to 32 bits before the compare, matching the 32-bit futex-word contract.
+    if current != (val as u32) {
         return FutexWaitOutcome::ValueMismatch;
     }
